Fix sidebar per route
